Found alot more Wallgen code, xD, moved to legacy
